Chnaged price long to float
git-svn-id: http://192.168.2.90/svn/bartsy-server/bartsy-server@387 c3b44e55-a100-42f6-9c86-32dbf579054b

+++ grails-app/domain/bartsy/Cocktails.groovy
@@ -29,7 +29,7 @@ class Cocktails {
 	String glass
 	String ingredients
 	String instructions
-	long price
+	float price
 	String available
 	Date dateCreated
 	Date lastUpdated
